refactor: curve out separateByAvailability
