+++ packages/textlint/src/util/find-util.ts
@@ -5,6 +5,7 @@ import glob from "glob";
 import path from "path";
 import fs from "fs";
 import debug0 from "debug";
+
 const debug = debug0("textlint:find-util");
 const DEFAULT_IGNORE_PATTERNS = Object.freeze(["**/.git/**", "**/node_modules/**"]);
 export type FindFilesOptions = {
@@ -91,35 +92,4 @@ export function findFiles(patterns: string[], options: FindFilesOptions = {}): s
         }
     });
     return files;
-}
-
-/**
- * classify files by extname
- * unAvailableFilePath will be used for warning.
- * If The user can use glob pattern like `src/*.js` as arguments, It will be expanded by shell.
- * pathsToGlobPatterns not modified that pattern.
- * So, unAvailableFilePath should be ignored silently.
- * @param {string[]} files
- * @param {{extensions?: string[]}} [options]
- * @returns {{availableFiles: string[], unAvailableFiles: string[]}}
- */
-export function separateByAvailability(
-    files: string[],
-    options: { extensions?: string[] } = {}
-): { availableFiles: string[]; unAvailableFiles: string[] } {
-    const extensions = options.extensions || [];
-    const availableFiles: string[] = [];
-    const unAvailableFiles: string[] = [];
-    files.forEach((filePath) => {
-        const extname = path.extname(filePath) || path.basename(filePath);
-        if (extensions.indexOf(extname) === -1) {
-            unAvailableFiles.push(filePath);
-        } else {
-            availableFiles.push(filePath);
-        }
-    });
-    return {
-        availableFiles,
-        unAvailableFiles
-    };
 }
